backend/requirements.txt
dj_rest_auth==6.0.0
Django==5.0.6
django_allauth==0.61.1
djangorestframework==3.15.1
drf_yasg==1.21.7
gunicorn>=20.1.0
django-filter
mysqlclient===2.2.4
djangorestframework-simplejwt
whitenoise
django-minio-storage